requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Language detection + translation
langdetect>=1.0.9
deep-translator>=1.11.4

# Embeddings + vector search
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# LLM client
openai>=1.30.0
anthropic>=0.28.0          # optional, swap in if using Claude

# Geo distance
geopy>=2.4.0
requests>=2.31.0

# Text normalization
Unidecode>=1.3.8           # for transliteration
rapidfuzz>=3.9.0           # fuzzy string matching (useful for Level 1)

# Testing
pytest>=8.0.0

# Notebook support (optional but nice to have)
jupyter>=1.0.0
